refactor: tests case
