+++ packages/element3/src/utils/popup1/index.js
@@ -1,3 +1,3 @@
-import PopupComponent from './PopupComponent.jsx'
+import PopupComponent from './PopupComponent.js'
 // import './style.scss'
 export { PopupComponent }
